backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
pydantic==2.6.1
httpx==0.26.0
feedparser==6.0.11
python-dotenv==1.0.1
python-multipart==0.0.9
